# src/quads/server/swagger.yaml
openapi: 3.0.0
info:
  version: '2.2.6'
  title: 'QUADS'
  description: |-
    QUADS automates the future scheduling, end-to-end provisioning and delivery of bare-metal servers and networks.

    This is it's REST API implementation.

    Some useful links:
    - [Quads repository](https://github.com/quadsproject/quads)
    - [Quads blog](https://quads.dev)
  license:
    name: GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: 'https://quads.example.com/api/v3/'
...
paths:
...
  /vlans/:
    get:
      summary: Returns a list with all the defined vlans
      tags:
        - Vlans
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Vlan'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Add a new vlan
      tags:
        - Vlans
      requestBody:
        description: Body request with *vlan* details
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - gateway
                - ip_free
                - ip_range
                - netmask
                - vlan_id
              properties:
                gateway:
                  type: string
                  example: 10.1.18.8
                ip_free:
                  type: integer
                  example: 256
                ip_range:
                  type: string
                netmask:
                  type: string
                  example: 255.255.255.255
                vlan_id:
                  type: integer
                  example: 601
          text/plain:
            schema:
              type: string
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
      security:
        - BearerAuth: [ ]
  /vlans/free:
    get:
      summary: Returns a list with all the free vlans
      tags:
        - Vlans
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Vlan'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /vlans/{vlan_id}:
    get:
      summary: Returns a detail of a specific vlan by id
      tags:
        - Vlans
      parameters:
        - name: vlan_id
          in: path
          description: Vlan ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Vlan details list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vlan'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
        summary: Update an existing vlan
        tags:
            - Vlans
        parameters:
            - name: vlan_id
              in: path
              description: Vlan ID
              required: true
              schema:
                type: integer
        requestBody:
          description: Body request with *vlan* details
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vlan'
        responses:
          '200':
            description: Vlan details
            headers:
              x-next:
                description: A link to the next page of responses
                schema:
                  type: string
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Vlan'
          default:
            description: Unexpected error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
        security:
            - BearerAuth: [ ]
    delete:
      summary: Delete vlan by vlan id
      tags:
        - Vlans
      parameters:
        - name: vlan_id
          in: path
          description: Vlan ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted Vlan
      security:
        - BearerAuth: [ ]
...
components:
  schemas:
...
    Vlan:
      type: object
      properties:
        gateway:
          type: string
          example: 10.1.18.8
        ip_free:
          type: integer
          example: 256
        ip_range:
          type: string
        netmask:
          type: string
          example: 255.255.255.255
        vlan_id:
          type: integer
          example: 601
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          example: 400
          format: int32
        message:
          type: string
          example: Something went horribly wrong.
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer